requirements.txt
coverage==4.4.2
Flask==0.12.2
Flask-DebugToolbar==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
psycopg2==2.7.3.2
PyYAML==3.12
GeoAlchemy2==0.4.2
imposm.parser==1.0.7
voluptuous==0.10.5
geojson==2.3.0
shapely==1.6.4